Fixed MsgBitfield being overwritten by recv()
Since we were passing a slice to the bitfield and bitfield message constructors, we are re-using the memory that is being recv()d into. This causes bitfields to be unintentionally overwritten on subsequent DecodeAll() calls.

Fixed by making a copy of the data in the MsgBitfield decode, and passing that to the Bitfield constructor